Workshop/requirements.txt
aiohttp
aiortc
numpy
pyserial
rpi_ws281x
adafruit-circuitpython-neopixel
adafruit-blinka
